+++ src/components/ReactPlayer/ReactPlayer.tsx
@@ -9,7 +9,7 @@ import React, {
     useRef,
     useState,
 } from 'react';
-import {Icon} from '@yandex-data-ui/common';
+import {Icon} from '@yandex-cloud/uikit';
 import {ClassNameProps} from '@yandex-data-ui/cloud-components';
 
 import {block} from '../../utils';