+++ src/context/mobileAppContext/constants.ts
@@ -1,5 +1,0 @@
-export enum Platform {
-    IOS = 'ios',
-    ANDROID = 'android',
-    BROWSER = 'browser',
-}